refactor: change function name auth.ts to getUserIdFromRequest.ts
