+++ src/pages/api/v1/product/delete/selected/index.ts
@@ -1,5 +1,5 @@
-import getUserIdFromRequest from "@/lib/auth";
 import { InternalServerError } from "@/lib/CustomErrors";
+import getUserIdFromRequest from "@/lib/getUserIdFromRequest";
 import ShoppingListRepository from "@/repository/ShoppingListRepository";
 import UserRepository from "@/repository/UserRepository";
 import ShoppingListService from "@/services/ShoppingListService";
